+++ blockfile/f_def.h
@@ -6,7 +6,7 @@
 
 //----------------------------------------------------------
 
-#define BFHEAD_LENGTH (sizeof(int)*2)    //file header size
+#define BFHEAD_LENGTH ((int)sizeof(int)*2)    //file header size
 
 #define TRUE 1
 #define FALSE 0
